refactor(scheduler): make the trigger registry real, and fix the 0045 downgrade
Review of the first pass found the abstraction was still shaped around
the original two triggers: window_hours was hard-coded into the generic
counter signature, edge model, serializer and alert template, and a
closed enum gated the kinds. 'Adding a third trigger touches the trigger
list and nothing else' was not true.

Extract the subsystem out of runtime.py (993 -> 592 lines) into
failure_guard.py, where each trigger owns its configuration, evaluation,
public details, alert text and reset predicate, and the engine knows
only how to load latches, detect crossings and collect edges. The proof
is a test registering a runtime-duration trigger with no count/window
contract at all, flowing through evaluation, persistence, serialization,
health and delivery without touching engine code.

Migration 0045's downgrade recreated the two legacy columns but never
dropped the normalized table, leaving a downgraded database matching
neither revision. It now restores the columns, drops the table, and
raises on a trigger kind it cannot represent rather than silently
dropping the row. All four latch states are covered exactly.
